+++ packages/api/lambdas/message-consumer.js
@@ -141,7 +141,7 @@ async function processRecord(record, fromSNS, enabledRules) {
         ...lookupCollectionInEvent(eventObject),
         sourceArn: get(parsed, 'eventSourceARN'),
         provider: lookupProviderInEvent(eventObject),
-        enableMismatch: process.env.enableMismatch,
+        allowProviderMismatchOnRuleFilter: process.env.allow_provider_mismatch_on_rule_filter,
       };
     } catch (error) {
       log.error('Caught error parsing JSON:', error);