+++ tf-modules/ingest/variables.tf
@@ -64,10 +64,10 @@ variable "default_s3_multipart_chunksize_mb" {
   default = 256
 }
 
-variable "enable_mismatch" {
+variable "allow_provider_mismatch_on_rule_filter" {
   description = "optional variable to be used in message_consumer lambdas for disabling rule/message provider mismatches"
   type = bool
-  default = true
+  default = false
 }
 
 variable "distribution_url" {